+++ django_cassandra_engine/models.py
@@ -1,0 +1,4 @@
+from django.db import connection
+
+
+connection.connect()  # connect at startup